refactor: update fetchPlaceholders import paths to use aem.js
