+++ scripts/initializers/auth.js
@@ -2,7 +2,7 @@
 import { initializers } from '@dropins/tools/initializer.js';
 import { initialize, setFetchGraphQlHeaders } from '@dropins/storefront-auth/api.js';
 import { initializeDropin } from './index.js';
-import { fetchPlaceholders } from '../scripts.js';
+import { fetchPlaceholders } from '../aem.js';
 import { getHeaders } from '../configs.js';
 
 await initializeDropin(async () => {
